requirements.txt
pillow~=10.2.0
flask~=3.0.0